requirements.txt
django==3.2.14
environs[django]==9.3.5
psycopg2-binary==2.9.2
django-phonenumber-field[phonenumbers]==6.3.0
django-timezone-field==5.0
tzdata==2022.1
djangorestframework==3.13.1
drf-yasg==1.21.3
celery==5.2.7
redis==4.3.4
requests==2.28.1
